navbar, some css, and renaming
