+++ react-frontend/src/components/BarberAvail/Table.jsx
