+++ react-frontend/src/components/navbar/navbar.css
@@ -1,0 +1,25 @@
+.nav_bar {
+  align-items: center;
+  justify-content: left;
+  list-style-type: none;
+  display: flex;
+  flex-direction: row;
+  color: silver;
+  padding: 0px;
+  margin-top: -10px;
+  margin-bottom: 0px;
+  background-color: #FBFBFF;
+  border-bottom: 4px solid #3A405A;
+  width: 106%;
+  height: 4%;
+  bottom: 0px;
+}
+
+.title {
+  color: #3A405A;
+}
+
+#navlink {
+  padding-right: 20px;
+  padding-left: 20px;
+}
